# swagger.yaml
openapi: 3.1.0
info:
  title: Paradise Cakes API
  version: 1.0.0
paths:
...
  /desserts/{dessert_id}:
    get:
      summary: Get Dessert
      operationId: get_dessert_desserts__dessert_id__get
      parameters:
        - name: dessert_id
          in: path
          required: true
          schema:
            type: string
            title: Dessert Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDessertResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    patch:
      summary: Patch Dessert
      operationId: patch_dessert_desserts__dessert_id__patch
      parameters:
        - name: dessert_id
          in: path
          required: true
          schema:
            type: string
            title: Dessert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchDessertRequest"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchDessertResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Dessert
      operationId: delete_dessert_desserts__dessert_id__delete
      parameters:
        - name: dessert_id
          in: path
          required: true
          schema:
            type: string
            title: Dessert Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDessertResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    DeleteDessertResponse:
      properties:
        dessert_id:
          type: string
          title: Dessert Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        prices:
          items:
            $ref: "#/components/schemas/Price"
          type: array
          title: Prices
        dessert_type:
          type: string
          title: Dessert Type
        ingredients:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Ingredients
        created_at:
          type: integer
          title: Created At
        last_updated_at:
          type: integer
          title: Last Updated At
        images:
          anyOf:
            - items:
                $ref: "#/components/schemas/Image"
              type: array
            - type: "null"
          title: Images
        visible:
          type: boolean
          title: Visible
          default: false
      type: object
      title: DeleteDessertResponse
...
    GetDessertResponse:
      properties:
        dessert_id:
          type: string
          title: Dessert Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        prices:
          items:
            $ref: "#/components/schemas/Price"
          type: array
          title: Prices
        dessert_type:
          type: string
          title: Dessert Type
        ingredients:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Ingredients
        created_at:
          type: integer
          title: Created At
        last_updated_at:
          type: integer
          title: Last Updated At
        images:
          anyOf:
            - items:
                $ref: "#/components/schemas/Image"
              type: array
            - type: "null"
          title: Images
        visible:
          type: boolean
          title: Visible
          default: false
      type: object
      title: GetDessertResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Image:
      properties:
        image_id:
          type: string
          title: Image Id
        url:
          type: string
          title: Url
        position:
          type: integer
          title: Position
        file_type:
          type: string
          title: File Type
      type: object
      required:
        - image_id
        - url
        - position
        - file_type
      title: Image
    PatchDessertRequest:
      properties:
        name:
          anyOf:
            - type: string
            - type: "null"
          title: Name
        description:
          anyOf:
            - type: string
            - type: "null"
          title: Description
        prices:
          anyOf:
            - items:
                $ref: "#/components/schemas/Price"
              type: array
            - type: "null"
          title: Prices
        dessert_type:
          anyOf:
            - type: string
            - type: "null"
          title: Dessert Type
        ingredients:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Ingredients
        images:
          anyOf:
            - items:
                $ref: "#/components/schemas/Image"
              type: array
            - type: "null"
          title: Images
        visible:
          anyOf:
            - type: boolean
            - type: "null"
          title: Visible
      type: object
      title: PatchDessertRequest
    PatchDessertResponse:
      properties:
        dessert_id:
          type: string
          title: Dessert Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        prices:
          items:
            $ref: "#/components/schemas/Price"
          type: array
          title: Prices
        dessert_type:
          type: string
          title: Dessert Type
        ingredients:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Ingredients
        created_at:
          type: integer
          title: Created At
        last_updated_at:
          type: integer
          title: Last Updated At
        images:
          anyOf:
            - items:
                $ref: "#/components/schemas/Image"
              type: array
            - type: "null"
          title: Images
        visible:
          type: boolean
          title: Visible
          default: false
      type: object
      title: PatchDessertResponse
...
    Price:
      properties:
        size:
          type: string
          title: Size
        base:
          type: number
          title: Base
      type: object
      required:
        - size
        - base
      title: Price
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError